[carla] New bounding boxes example script (#9179) * bboxes iteration1
* added light state to JSON

This PR adds a new script to show how to manage bounding boxes for actors within a CARLA simulation and also to project them into the camera plane and derive 2D bounding boxes from the instance segmentation images. The script can also record 2D and 3D bounding boxes for each frame in a JSON file alongside a PNG of the RGB camera view for the corresponding frame.